refactor: move useOpenStatus hook to main.ts file
